refactor(components): extract constructable stylesheet probe into own module | rk | #4092
