+++ packages/components/src/utils/applyConstructableStylesheetStyle.ts
@@ -14,7 +14,7 @@ const getElementMap = (element: HTMLElement): ElementMap => {
   return elementsMap.get(tagName);
 };
 
-// TODO: Use function from ./jss (Causes bundling issues)
+// TODO: Use function from ./supportsConstructableStylesheets (Causes bundling issues)
 const hasConstructableStylesheetSupport = ((): boolean => {
   try {
     return typeof new CSSStyleSheet().replaceSync === 'function';
